+++ src/arclet/alconna/_internal/_analyser.py
@@ -27,10 +27,10 @@ from ..model import HeadResult, OptionResult, SubcommandResult
 from ..output import output_manager
 from ..typing import TDC, InnerShortcutArgs
 from ._handlers import (
+    HEAD_HANDLES,
     _handle_shortcut_data,
     _handle_shortcut_reg,
     analyse_args,
-    analyse_header,
     analyse_param,
     handle_completion,
     handle_help,
@@ -171,6 +171,7 @@ class SubAnalyser(Generic[TDC]):
         if not trigger:
             name, _ = argv.next(sub.separators)
             if name not in sub.aliases:
+                argv.rollback(name)
                 if not argv.fuzzy_match:
                     raise InvalidParam(lang.require("subcommand", "name_error").format(source=sub.dest, target=name))
                 for al in sub.aliases:
@@ -224,6 +225,8 @@ class Analyser(SubAnalyser[TDC], Generic[TDC]):
     """命令实例"""
     command_header: Header
     """命令头部"""
+    header_handler: Callable[[Header, Argv], HeadResult]
+    """头部处理器"""
 
     def __init__(self, alconna: Alconna[TDC], compiler: TCompile | None = None):
         """初始化解析器
@@ -238,6 +241,7 @@ class Analyser(SubAnalyser[TDC], Generic[TDC]):
     def compile(self, param_ids: set[str]):
         self.extra_allow = not self.command.meta.strict or not self.command.namespace_config.strict
         self.command_header = Header.generate(self.command.command, self.command.prefixes, self.command.meta.compact)
+        self.header_handler = HEAD_HANDLES[self.command_header.flag]
         self._compiler(self, param_ids)
         return self
 
@@ -301,7 +305,7 @@ class Analyser(SubAnalyser[TDC], Generic[TDC]):
             ArgumentMissing: 参数缺失
         """
         try:
-            self.header_result = analyse_header(self.command_header, argv)
+            self.header_result = self.header_handler(self.command_header, argv)
         except InvalidParam as e:
             _next = e.args[1]
             if _next.__class__ is not str or not _next: