+++ CHANGELOG.md
@@ -1,5 +1,58 @@
 # 更新日志
 
+## Alconna 1.8.10
+
+### 改进
+
+- 适配 `NEPattern` 0.7.0
+
+## Alconna 1.8.9
+
+### 修复
+
+- 修复传入命令首个元素非字符串时能错误触发快捷指令的问题
+
+## Alconna 1.8.8
+
+### 修复
+
+- 修复子命令在自身名称解析失败时其结果被错误存储的问题
+
+## Alconna 1.8.7
+
+### 改进
+
+- 头部解析逻辑拆分
+
+### 修复
+
+- 修复 manager 更新命令时未使用 `Argv` 的 cache 问题
+
+## Alconna 1.8.6
+
+### 改进
+
+- 帮助信息现在会显示命令的快捷指令
+
+### 新增
+
+- `Alconna.shortcut` 新增参数 `humanized`，用于指定快捷指令的可读性文本以供显示
+
+    ```python
+    from arclet.alconna import Alconna, Args
+  
+    alc = Alconna("test", Args["foo", int])
+    alc.shortcut("t(\d+)", command="test {0}", humanized="t[数字]")
+    ```
+
+- `CommandMeta` 新增属性 `hide_shortcut`, 用于在帮助信息里隐藏命令的快捷指令
+
+## Alconna 1.8.5
+
+### 修复
+
+- 修复 `Alconna` 的 aliases (继承自subcommand的) 错误
+
 ## Alconna 1.8.4
 
 ### 新增